minor enhancement

+++ texture/checker.cuh
@@ -17,9 +17,9 @@ class Checker : public Texture {
 
 __host__ __device__
 Arr3 Checker::map(float u, float v, const Arr3 &point) const {
-  auto sines = sin(10.0f * point.x()) * sin(10.0f * point.y()) * sin(10.0f * point.z());
+  auto sines = sinf(10.0f * point.x()) * sinf(10.0f * point.y()) * sinf(10.0f * point.z());
 
-  if (sines < 0) {
+  if (sines < 0.0f) {
     return this->odd->map(u, v, point);
   }
