+++ utility/camera.cuh
@@ -31,7 +31,7 @@ Camera::Camera(
   float focusDist,float time0, float time1
 ) {
 	double theta = degreesToRadians(vfov);
-	double height = tan(theta / 2);
+	double height = tanf(theta / 2.0f);
 	double viewportHeight = 2.0 * height;
 	double viewportWidth = aspectRatio * viewportHeight;
 
@@ -42,9 +42,9 @@ Camera::Camera(
 	this->origin = lookfrom;
 	this->horizontal = focusDist * viewportWidth * this->u;
 	this->vertical = focusDist * viewportHeight * this->v;
-	this->lowerLeftCorner = origin - horizontal / 2 - vertical / 2 - focusDist * this->w;
+	this->lowerLeftCorner = origin - horizontal / 2.0f - vertical / 2.0f - focusDist * this->w;
 
-	this->lensRadius = aperture / 2;
+	this->lensRadius = aperture / 2.0f;
 }
 
 __device__
